feat(Next > Today): Album Page and Today display the same view
